CCUI-2972 #comment Simplified implementation with a hook
