+++ packages/rapid-cmi5/src/lib/design-tools/rapidcmi5_mdx/toolbar/components/InsertQuotes.tsx
@@ -30,6 +30,7 @@ import {
 } from '@rapid-cmi5/ui';
 import { MUIButtonWithTooltip } from './MUIButtonWithTooltip';
 import { useCallback, useState } from 'react';
+import { useSelectionHelper } from 'packages/rapid-cmi5/src/lib/hooks/useSelectionHelper';
 
 /**
  * A toolbar button component that inserts a quotes into the editor.
@@ -40,6 +41,7 @@ export const InsertQuotes = ({ isDrawer }: { isDrawer?: boolean }) => {
   const editor = useCellValue(rootEditor$) as LexicalEditor | null;
   const [syntaxExtensions] = useCellValues(syntaxExtensions$);
   const theme: any = useTheme();
+  const selectionHelper = useSelectionHelper();
   const [isConfiguring, setIsConfiguring] = useState(false);
 
   /**
@@ -56,15 +58,7 @@ export const InsertQuotes = ({ isDrawer }: { isDrawer?: boolean }) => {
         editor.update(() => {
           let selection = $getSelection();
           if (!$isRangeSelection(selection) || !selection.isCollapsed()) return;
-
-          const anchorNode = selection.anchor.getNode();
-          const topLevel = anchorNode.getTopLevelElement();
-          if (topLevel && topLevel.getType() !== 'paragraph') {
-            const paragraph = $createParagraphNode();
-            topLevel.insertAfter(paragraph);
-            paragraph.select();
-            selection = $getSelection() as typeof selection;
-          }
+          selection = selectionHelper.getInsertSelection(selection);
 
           // create child quotes content
           const theChildMDast = convertMarkdownToMdast(
