+++ packages/rapid-cmi5/src/lib/design-tools/rapidcmi5_mdx/toolbar/components/InsertStatements.tsx
@@ -24,6 +24,7 @@ import {
 } from '@rapid-cmi5/ui';
 import { MUIButtonWithTooltip } from './MUIButtonWithTooltip';
 import { useCallback, useState } from 'react';
+import { useSelectionHelper } from 'packages/rapid-cmi5/src/lib/hooks/useSelectionHelper';
 
 /**
  * A toolbar button component that inserts a statements block into the editor.
@@ -34,6 +35,7 @@ export const InsertStatements = ({ isDrawer }: { isDrawer?: boolean }) => {
   const editor = useCellValue(rootEditor$) as LexicalEditor | null;
   const [syntaxExtensions] = useCellValues(syntaxExtensions$);
   const theme: any = useTheme();
+  const selectionHelper = useSelectionHelper();
   const [isConfiguring, setIsConfiguring] = useState(false);
 
   /**
@@ -50,14 +52,7 @@ export const InsertStatements = ({ isDrawer }: { isDrawer?: boolean }) => {
           let selection = $getSelection();
           if (!$isRangeSelection(selection) || !selection.isCollapsed()) return;
 
-          const anchorNode = selection.anchor.getNode();
-          const topLevel = anchorNode.getTopLevelElement();
-          if (topLevel && topLevel.getType() !== 'paragraph') {
-            const paragraph = $createParagraphNode();
-            topLevel.insertAfter(paragraph);
-            paragraph.select();
-            selection = $getSelection() as typeof selection;
-          }
+          selection = selectionHelper.getInsertSelection(selection);
 
           // create statements node
           const mdastStatements: ContainerDirective = {
